Cleanup code

+++ src/nrf_802154.c
@@ -309,7 +309,8 @@ void nrf_802154_antenna_config_set(nrf_802154_ant_div_config_t config)
     nrf_802154_ant_div_config_set(config);
 
 }
-void nrf_802154_ant_div_toggle_time_set(uint32_t toggle_time)
+
+void nrf_802154_ant_div_toggle_time_set(uint8_t toggle_time)
 {
     nrf_802154_pib_ant_div_toggle_time_set(toggle_time);
 }
