+++ src/nrf_802154_pib.h
@@ -368,7 +368,7 @@ nrf_802154_ant_div_antenna_t nrf_802154_pib_ant_div_antenna_get();
  *
  * @param[in] toggle_time  Time between switching antennas in us.
  */
-void nrf_802154_pib_ant_div_toggle_time_set(uint32_t toggle_time);
+void nrf_802154_pib_ant_div_toggle_time_set(uint8_t toggle_time);
 
 /**
  * @brief Reads the current antenna toggle time in automatic mode.
